Create Models
This code creates a Person class with two properties (Name and Age) and an Introduce() method that prints that person's representation to the console. In the Main() method, a list of people is created, populated with instances of the Person class, and then each person in the list is represented.